Basuru's Freelancer & Client pages re designed
